chore: move interface of socket class
